Move Plugin and hooks to API
