+++ plugin-api/src/main/java/net/clgd/ccemux/api/plugins/hooks/RendererCreated.java
@@ -1,4 +1,4 @@
-package net.clgd.ccemux.plugins.hooks;
+package net.clgd.ccemux.api.plugins.hooks;
 
 import net.clgd.ccemux.api.emulation.Emulator;
 import net.clgd.ccemux.api.rendering.Renderer;
