+++ plugin-api/src/main/java/net/clgd/ccemux/api/plugins/hooks/package-info.java
@@ -1,0 +1,8 @@
+
+/**
+ * This package contains various hooks which can be implemented by CCEmuX
+ * plugins to provide functionality. A list of common hooks can be
+ * found {@link net.clgd.ccemux.api.plugins.hooks.Hook here}.
+ * @see net.clgd.ccemux.api.plugins.hooks.Hook
+ */
+package net.clgd.ccemux.api.plugins.hooks;
