+++ src/main/java/net/clgd/ccemux/plugins/builtin/AWTPlugin.java
@@ -13,9 +13,9 @@ import net.clgd.ccemux.api.config.ConfigProperty;
 import net.clgd.ccemux.api.config.Group;
 import net.clgd.ccemux.api.emulation.EmuConfig;
 import net.clgd.ccemux.api.emulation.EmulatedComputer;
+import net.clgd.ccemux.api.plugins.Plugin;
 import net.clgd.ccemux.api.rendering.Renderer;
 import net.clgd.ccemux.api.rendering.RendererFactory;
-import net.clgd.ccemux.plugins.Plugin;
 import net.clgd.ccemux.rendering.awt.AWTRenderer;
 import net.clgd.ccemux.rendering.awt.config.ConfigView;
