+++ src/main/java/net/clgd/ccemux/plugins/builtin/JFXPlugin.java
@@ -10,8 +10,8 @@ import net.clgd.ccemux.api.OperatingSystem;
 import net.clgd.ccemux.api.config.ConfigProperty;
 import net.clgd.ccemux.api.config.Group;
 import net.clgd.ccemux.api.emulation.EmuConfig;
+import net.clgd.ccemux.api.plugins.Plugin;
 import net.clgd.ccemux.api.rendering.RendererFactory;
-import net.clgd.ccemux.plugins.Plugin;
 import net.clgd.ccemux.rendering.javafx.JFXRendererFactory;
 
 @AutoService(Plugin.class)
