+++ src/main/java/net/clgd/ccemux/plugins/builtin/TRoRPlugin.java
@@ -7,8 +7,8 @@ import java.util.Optional;
 import com.google.auto.service.AutoService;
 
 import net.clgd.ccemux.api.emulation.EmuConfig;
+import net.clgd.ccemux.api.plugins.Plugin;
 import net.clgd.ccemux.api.rendering.RendererFactory;
-import net.clgd.ccemux.plugins.Plugin;
 import net.clgd.ccemux.rendering.tror.TRoRRenderer;
 
 @AutoService(Plugin.class)
