--- src/main/java/net/clgd/ccemux/plugins/hooks/package-info.java
@@ -1,8 +1,0 @@
-
-/**
- * This package contains various hooks which can be implemented by CCEmuX
- * plugins to provide functionality. A list of common hooks can be
- * found {@link net.clgd.ccemux.plugins.hooks.Hook here}.
- * @see net.clgd.ccemux.plugins.hooks.Hook
- */
-package net.clgd.ccemux.plugins.hooks;
